feat: split source files into dirs
