+++ src/build-nc-w64/helapordo_win.c
