+++ src/build-nc/game_curses.c
