+++ src/build-nc/helapordo.c
