+++ src/core/equips.c
