+++ src/core/equips.h
