+++ src/core/game_core.c
