+++ src/core/game_lore.c
