+++ src/core/game_lore.h
