+++ src/core/game_lore_alt.c
