+++ src/core/sprites.c
