+++ src/core/sprites.h
