--- src/game_animations.h
@@ -1,25 +1,0 @@
-#ifndef GAME_ANIMATIONS_H_
-#define GAME_ANIMATIONS_H_
-#include "./animations/knight_tapis.h"
-#include "./animations/mage_spark.h"
-#include "./animations/archer_drop.h"
-#include "./animations/assassin_poof.h"
-
-#include "./animations/mummy_shuffle.h"
-#include "./animations/ghost_spell.h"
-#include "./animations/boar_scream.h"
-#include "./animations/troll_club.h"
-#include "./animations/goblin_shoot.h"
-#include "./animations/zombie_walk.h"
-#include "./animations/imp_fireball.h"
-#include "./animations/werewolf_transform.h"
-
-#include "./animations/crawlingdude_crawl.h"
-#include "./animations/srwarthog_square.h"
-#include "./animations/headlessninja_throw.h"
-#include "./animations/bluetroll_wonder.h"
-
-#include "./animations/enter_door.h"
-#include "./animations/alt_chest_opening.h"
-
-#endif // GAME_ANIMATIONS_H_
