+++ src/utils/game_debug.c
