+++ src/utils/game_debug.h
