+++ src/utils/game_fight.c
