+++ src/utils/game_init.c
