+++ src/utils/game_utils.c
