+++ src/utils/rooms.c
