+++ src/utils/saves.h
@@ -19,16 +19,16 @@
 #ifndef SAVES_H_
 #define SAVES_H_
 
-#include "equips.h"
+#include "../core/equips.h"
 #include "game_init.h"
 
 #ifdef HELAPORDO_CURSES_BUILD
-#include "game_curses.h"
+#include "../build-nc/game_curses.h"
 #else
 #ifndef HELAPORDO_RAYLIB_BUILD
 #error "HELAPORDO_CURSES_BUILD and HELAPORDO_RAYLIB_BUILD are both undefined.\n"
 #else
-#include "game_rl.h"
+#include "../build-rl/game_rl.h"
 #endif // HELAPORDO_RAYLIB_BUILD
 #endif // HELAPORDO_CURSES_BUILD
